+++ container/mock.go
@@ -15,7 +15,7 @@
 package container
 
 import (
-	info "github.com/google/cadvisor/info/v1"
+	info "github.com/Clever/cadvisor/info/v1"
 	"github.com/stretchr/testify/mock"
 )