+++ manager/manager_test.go
@@ -22,12 +22,13 @@ import (
 	"testing"
 	"time"
 
-	"github.com/google/cadvisor/container"
-	"github.com/google/cadvisor/container/docker"
-	info "github.com/google/cadvisor/info/v1"
-	itest "github.com/google/cadvisor/info/v1/test"
-	"github.com/google/cadvisor/storage/memory"
-	"github.com/google/cadvisor/utils/sysfs/fakesysfs"
+	"github.com/Clever/cadvisor/collector"
+	"github.com/Clever/cadvisor/container"
+	"github.com/Clever/cadvisor/container/docker"
+	info "github.com/Clever/cadvisor/info/v1"
+	itest "github.com/Clever/cadvisor/info/v1/test"
+	"github.com/Clever/cadvisor/storage/memory"
+	"github.com/Clever/cadvisor/utils/sysfs/fakesysfs"
 )
 
 // TODO(vmarmol): Refactor these tests.
@@ -52,7 +53,7 @@ func createManagerAndAddContainers(
 			spec,
 			nil,
 		).Once()
-		cont, err := newContainerData(name, memoryStorage, mockHandler, nil, false)
+		cont, err := newContainerData(name, memoryStorage, mockHandler, nil, false, &collector.FakeCollectorManager{})
 		if err != nil {
 			t.Fatal(err)
 		}